test: rework funding logic: AcceptFunding and CreateFunding events.
The Funding object was weird.  Create two new events which stash
events, and it's much more like everything else.

This also creates a 'helpers.py' for the tests, with some useful
things in it, removing 'blocks.py'.

With this, we no longer make assumptions about the funding keys used
by the runner.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>