dual-funding: add another utxo the common tx_spendable
Adds another output to `tx_spendable` which is locked to a 1of5 multisig
output.

This will allow us to test underpayment of witness fees for the v2
channel open protocol.

Helpful notes for updating this tx in the future.

 - The block submitted to bitcoind at startup pays to a coinbase with the
listed privkey (checked in in this PR).
 - tx_spendable spends this coinbase output.

To create a new/updated tx_spendable, the following should be run
against a fresh bitcoind (on regtest):

    bitcoin-cli submitblock <block>
    bitcoin-cli importprivkey <provided privkey>
    bitcoin-cli generatetoaddress 100 <some random address>
    bitcoin-cli signrawtransactionwithwallet <new candidate tx_spendable>

The resulting tx_spendable can be swapped out. Don't forget to also
update the notes explaining the outputs.