Give the possibility to run lnprototest with different version of bitcoin core.
The object of this commit is to give the possibility to run different the test with different version of Bitcoin core without put change to the code (when it is not necessary). However, the object of this commit it is not to support very old version of bitcoin core, but only the version around the actual official version of it.

Rebased on top of the master

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>