Update dependence and remove pyln.testing dependences
As mentioned correctly in issue #17 the pyln.testing require the postgress dependencies, but inside the lnprototest there are no any reason to use the package pyln.testing, because is used only util functions. With this commit, I removed the dependencies, and I reproduce the few lines of code that are imported from the pyln package.

Changelog: Remove pyln.testing to remove the dependencies problem with pip

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>