bitcoind backend: assign port at start time, not init.
I found a similar race in our core-lightning CI: when a node restarts,
there's a race where another node could start on the same port.

In this case, we catch the exception, which was covering up the failure.
The test then talks to the wrong bitcoind, which has unpredictable results.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>