cln: Use full hostname:port `rpcconnect` instead of just `rpcport`
I was having some difficulty running lnprototest locally because it
was reading my `bitcoin.conf` which points to a remote `bitcoind` that
doesn't have `regtest` running. This caused tests to hang and time out
since it couldn't connect to `bitcoind`. This now specifies that we
want to connect to the local regtest node we just spun up.