fix: bisect the test init
Allow bisecting the init test to make the debug
of some failure easier.

However, this is not a solution, because the problem
is when a single test needs to follow multiple branches
with the `TryAll` clause.

In this case, lnprototest can run a branch where the connection
needs to be aborted and then will run a branch that required
that the peer is connected, so at this point, lnprototest fails
with the stack trace reported in https://github.com/rustyrussell/lnprototest/issues/66

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>