fix: add Wait event to allow to slow down lnptototest
There is some case where lnprototest send the message too fast, well
lnprototest is always too fast but in most case, we simulate the `time.sleep`
inside the ExpectedMsg that iterates for a while till the message
is not received.

This is no longer true if lnprototest sends a message and the next steps
is not the ExpectedMsg or equal but a message that required that the
message sent to lnprototest is already processed, like in
the `FundChannel(...)` case.

So this commit is introducing a new Event `Wait` that
will allow us to wait some time before processing the next message.

This is needed when you send the connection event, and then
you send the fundchannel message. It is possible endup
in a concurrency problem when the node is waiting to finish
the connection, but receive the next message, just because lnprototest
it too fast.

All this to said that this is an hack (but an hack with docs :) )

Link: https://github.com/rustyrussell/lnprototest/issues/74
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>