feat: Enable stashing of expected messages
There are cases where we need to access the message that was
just decoded within the `ExpectMsg` event.

This commit adds a generic method to the stash module,
allowing retrieval of information from the message.

Furthermore, starting from this point, lnprototest will stash
all messages that are decoded using `ExpectMsg`. This enhancement
simplifies cross-compatibility efforts with other implementations.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>