fix: Grab the feature from the init message if it is not specified
This commit improves the method for generating the sequence
of events needed to connect to a node from the sender side.

Specifically, we can now retrieve the same feature
that the node sends to the runner.

This is useful because if the features do not match,
the connection must be closed.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>