feat: Open channel helper ignores announcement signatures
Signature exchanges are signed with the node's private key and verified by
the counterpart using the node's ID (which is the public key). The signatures
should not be the same because lightning implementation and lnprototest do not have the same
private key, or better yet, lnprototest will fake a different private key.

Unfortunately, we lack the necessary tools to easily create this signature,
which makes lnprototest a painful process. Apologies for the inconvenience!

Therefore, at some point, we should have a tool to calculate the signatures.
But for now, we are leveraging the lightning specification and
skipping the announce signature message.

This significantly simplifies the workflow.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>