feat: Make default runner connection available
There is no reason to keep a generic interface in the
runner and provide the implementation in the core lightning runner
if the only way to establish a connection with the node is through
the pyln.proto package.

This commit moves the CLightningConn inside the runner interface
to allow everyone access to the default implementation.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>