feat(test): Add a new test to ensure proper signature checking
Fixes a bug in lnprototest by removing the problematic
code outlined in patch [1].

During our investigation of the cln code, we discovered that
the message verification was not performed correctly as the BOL 7
suggest. This commit includes a patch [2] that fixes the issue and
introduces an integration test to validate that core lightning adheres
to the signature verification guidelines outlined in BOLT7.

[1] https://github.com/rustyrussell/lnprototest/pull/91
[2] https://github.com/ElementsProject/lightning/pull/6384

Reported-by: lnprototest (https://github.com/rustyrussell/lnprototest/pull/91)
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>