fix: Correct error check for expected errors
This commit addresses an issue with the error checking
mechanism. Currently, the `ExpectError` function only verifies
the occurrence of an error during the connection, without considering
the specific error message expected.

The upcoming changes in [1] will simplify lnprototest, but until then,
this fix provides a straightforward and effective solution for
checking error messages.

[1] https://github.com/rustyrussell/lnprototest/pull/95

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>