feat: Allowing the option to skip the creation of the wallet in bitcoind
A few years ago, I added this code for compatibility
with Bitcoin Core 21.

Now that I am integrating it with the LDK node runner, and LDK
can use a wallet to set up a node with Bitcoin Core, we need
a way to disable the wallet from Bitcoin Core to keep
our integration simpler.

Without this change, Bitcoin Core would have two wallets,
and we would need to set up the correct wallet in our architecture.

However, we can simply disable it to avoid this issue.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>