fix: Correct reorg issue in the ln helper function
Previously, an unexpected behavior in the ln helper function had gone
unnoticed. However, while Rusty Russell was working on [1], he
discovered a bug in lnprototest.

Specifically, within the helper function that opens a channel, there
was an incorrect way of counting the current block height. Ideally,
this logic should be abstracted elsewhere.

This commit addresses the block height counting
error and resolves the bug identified in [1].

[1] https://github.com/ElementsProject/lightning/pull/6628

Link: https://discord.com/channels/899980449231814676/941465665540325397/1154259454716563537
Reported-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>