events: throw an exception if the peer did not disconnect
Currently an event is too many things, and means too many things.

However, an event should return Ok or an error in most of the case.

If it is ok there is no error return, otherwise an exception. So
this commit is fixing a while loop while the event was returning always
false.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>