meta: fix cln version to 24.05
cln version 24.05 is the last version of core lightning that it is
working with the version of prototest. Most lickly this is a cln
bug described in the following issue [1].

However, we pin the cln version and report the bug to mainline probably
someone will be able to find the real issue in the cln or lnprototest
at some point.

[1] https://github.com/rustyrussell/lnprototest/issues/123
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>